allow users to submit new events at /submit-event
this commit adds a public view for submitting new events and adds two new
fields to the Event model:

- user_submitted :: a field that will allow us to track which events originated
from users and which originated from the admin backend interface

- is_approved :: all user-submitted events will default to being
unapproved. unapproved events will not show in the calendar.

this commit also allows the admin user to filter to only the unapproved events
to allow for simple moderation (by changing the is_approved field to true). an
actual moderation interface might make more sense if this feature actually gets
used.